+++ docs/app/components/content/examples/table/TableVirtualizeExample.vue
@@ -79,6 +79,7 @@ const columns: TableColumn<Payment>[] = [
 
 <template>
   <B24Table
+    sticky
     virtualize
     :data="data"
     :columns="columns"